Start conversion to API
Remove placeholder "index" and replace it with personalId.  Remove
references to sampleData and instead use a single GET to fetch all the
results.  This is very much not optimized at the moment.

I haven't done any refactoring yet, but we will probably want to do
that in future commits.

As of this commit, search works to fetch results via the API, but we
can't click on those results to view full results, nor can we edit or
add clients.  Coming soon!