Incomplete: PUT client data
As now written, this PUTs an incomplete client through to the server and
gets a 500 response.  I still need to figure out why that is, but wanted
to have this intermediate commit here so that I could return to this
state.  The good is that I'm successfully sending information, but the
bad is that it isn't updating the database correctly.