Working: PUT client data
As of this commit, I can PUT client data through and successfully change
the database.  I still need to make changes in index.js so that the
correct values are being passed (e.g., integers instead of strings for
race, ethnicity, and gender).  We're not saving gender yet, either.  All
of these should be trivial, though, now that PUT itself is working.