Implement improved buttons for intake #9
The intake form is now much user-friendlier in that different states
that are reached by having changed / not having changed data pertaining
to an old client / new client lead to different buttons that make sense
in context and protect the user from silly mistakes. The scheme is fully
described in
https://github.com/OpenTechStrategies/openhmis-intake/issues/9.

Additionally, make some changes to the template to hold information that
is stored in the DB in one form but presented to the user in a different
form. Prime examples include the DOB (stored as 'YYYY-MM-DD'; presented
as 'DD MMM YYYY' with the person's age attached) and picture (where the
image filename is distinct from the <img> tag src attribute, since the
latter includes a path to the image directory). Now the unformatted DB
versions are stored in hidden input fields in the form. This allows
streamlining of, for example, saving a new entity by pulling values from
<input> elements without having to build in exceptions to deal with DOB,
picture, etc.