Save only one copy of a new person.
Thanks to a bug introduced in a683576, PUT and POST requests were being
sent once per search keystroke (!).  Clearly that is incorrect behavior,
and this commit fixes it by moving the "keyup" call *inside* the
original GET, and closes the keyup so that it fires only searching, as
intended.  Previously all the on clicks were also happening inside that
function, which resulted in the multiple requests we saw earlier.