API: Save SSN, get closer to saving race
SSN now saves correctly for new and edited clients.  Race does not, in
part because I'm not sure what the API expects.  It looks like I should
be sending boolean values for each race, but then the db tables I have
only have room for one "Race Key."  Values for that race key are defined
in the OpenHMIS tree, as are boolean values like "amIndAKNative."  Which
should be stored?  If the latter, where do they show up in the db?  This
is a question for @slifty, probably.