API: Save and display race correctly
This commit saves and displays race correctly, with the exception of the
"client doesn't know" and "client refused" options.  I removed those
from the display in index.jade to more accurately reflect what I'm
storing at the moment.  I'm now sending the races as boolean values, as
is intended in the API.  I renamed the race checkboxes to the names of
the API properties, so that they would fit with the intake-form-filling
loop in index.js.  I added a missing comma in
client.server.controller.js.