Merge success: save, revert, & cancel buttons work
I've renamed the "SSN" input to lower case "ssn," so that it matches the
API element.  I may need to do the same thing with "DOB"/"dob".  I'm
going to test a little more first to make sure that change doesn't break
anything else.  Otherwise, this commit is mainly about giving new
functions the information they need, now that we're no longer relying on
a sample data set defined by a global variable.  In several places, I've
renamed "index" to "personalId" so that it matches the API, again.  I'm
afraid that this will need another, more careful sweep to make sure that
we aren't calling the same thing by different names.  I've had to define
"newEntity" and "origEntity" carefully in the revertChanges() and
checkForChanges() functions.

The next commit will test changing the date of birth in all situations.
That is also a bit of a special case, and will likely need some changes
to work correctly.