Send correct (empty) personalId for new client.
In the course of merging, I started sending "-1" as the personalId for a
newly created client, which resulted in them not being saved!  This
commit sets personalId to the empty string so that the db will create a
new id, as intended.

I also added a comment at the end of the wrapper function, which isn't
really related.