Fix #11: Send response to browser from PUT & POST
With many (many!) thanks to @slifty, here's what was going on: The
browser sent a request to the Node server, which passed it on to the
Java server and made the correct change.  All of that was as expected.
The Java server responded to the Node server, saying "Thanks, I got your
message."  However, the Node server was never sending anything back to
the browser to say that the request was finished.  The browser, then,
every two minutes sent another request to the Node server, which
faithfully passed it on to the Java server and soon we had two db
entries when we should have only had one.

I added a res.send() line in both the addClient() and editClient()
functions in `client.server.controller.js`.  This means that Node now
comes back to the browser and says "the request is finished, all is
well."  So, the browser will only send one request (as it should).