Start #12 and fix #17
`checkForChanges()` now works with select and checkbox elements, thanks
to the new on-change tests.  I added a `location.reload()` in
saveChanges(), so that users searching would always be seeing the latest
information (see #17).

We'll need to change that system for production use, though.  Pulling
all the data there is will be very slow once we're getting thousands of
records!