Fix #16: Handle special chars in names
@nttaylor made a regex change so that we accept special chars as long as
they don't have whitespace before or after them.  This works for cases
like "Jean-Marie" or "O'Malley."  It doesn't work for a case where the
special character is the first or last symbol in any part of the name.
It also doesn't work for names that have more than two parts separated
by whitespace.

Users will be able to edit the name on the intake form.  These concerns
are just about autofilling the name based on the search string.