Fix #21: Disable races if data not collected
I set the race checkboxes to be unchecked and disabled if "data not
collected" or "client refused" or "client doesn't know" are selected.
However, there is no way to save these options via the API, since the
races are all booleans and there are no variables corresponding to those
options (as far as I can tell).  So, the next time a user visits this
profile the "data not collected" or "client refused" or "client doesn't
know" checkbox will not be checked.  I'll open a separate issue about
that.

I also moved the "Data Not Collected" option to the top of the select
elements, so that it would be selected by default more easily.  I had
tried a couple other ways to do this but this was the simplest.

I'll try to check the change events for the checkboxes in a drier way in
future, rather than registering a separate callback for each one.  I
added the "single_checkbox" class for that reason.