Start #25: Read imported file into browser
Add a conditional to the controller to handle POST requests.  For now,
this reads the file into a variable and then sends it back to the
browser.  We'll do something different in the next commit.  Add a POST
route pointing to the controller.  Add a form to the front page to
receive the file for import (I'll have to move this, since it's right
above the search bar now).