Continue #25: Send file via AJAX
Instead of just submitting a form, I'm now sending the file to be
imported via AJAX so that we can interpret the data once it is returned
from the server, and send our POSTs via the API (for non-duplicate
rows).

Note that future commits will assume that we're importing a csv of a
format like "Client.csv" in the examples directory.  Does that sound
right?