Continue #25: Parse csv into array
As of this commit, the imported file is sitting in a JS array in the
browser.  We still need to check the SSN against the API for duplicates,
and to POST each non-duplicate record via the API.

I've done some of the parsing inside the controller, which may be bad
form.  I just split the csv so that the headers and footers aren't sent
back to the browser.  Then, in the client-client (as we're calling it),
I use the MIT-licensed "PapaParse" library to transform the csv lines
into a JS array.

I also fixed a call in index.jade so that it refers to PapaParse instead
of to a different jquery library that I decided not to use.

I added papaparse.js so that we could use the library.  It has a license
header and I don't think we need any other files from that repo.  They
do offer a minified version, if we prefer.