Improve HMIS/HUD compliance of export (#14).
Include the required Export.csv file in the export now.  Make
Client.csv be much closer to the format described in the HMIS CSV
Specifications (4_0FINAL).  Note that this means deviating from the
example provided in `examples/Client.csv`, particularly in the way the
Race columns are handled but also in other columns; I think that
example file is not fully standards-compliant, so deviation is okay.