Save pretty-printed JSON data for reference.
These are the JSON results of making the corresponding API requests
(e.g., http://localhost:8080/openhmis/services/clients, etc) against
the sample data from commit 55b298149 of the OpenHMIS server in
https://github.com/PCNI/OpenHMIS/blob/feature-compass_schema/.

An argument could be made that these files should really be stored in
that tree instead of here, if they're going to be stored anywhere.
But I need them as a reference for developing the client-side export,
so it's most convenient to have them here.  It's fine to remove them
once they're no longer relevant for development, though I'll probably
forget to do so.