Start import sorting: #25
Test whether a record to be imported has an SSN that matches an existing
record.  This commit adds several lines for interpreting the response
from the Node server.  Those may change.  I just tested with a file that
was exported from our db and PapaParse didn't split the lines into their
own arrays as it should, so it looked like the whole file was in one
array (which meant that the rest of the code did not work correctly).
Next commit will resolve that and possibly add the POST lines for new
data.  I need to turn the csv record into a correctly formatted Client,
before POSTing.