Working import: #25
Simplify the controller's split of file contents away from header and
footer.  I think that separating on `\r\n` will generally work for files
we create, but I worry somewhat that other files will have `\r\n` on
every line of data.  Add a div to hold notifications of duplicates from
the import (this will look nicer in the next commit.  Consider it
something of a placeholder for now).  Track duplicates in a variable
instead of just outputting them to the console.

Insert the data from each (non-duplicate) file line into a client
object.  I'm a little worried that explicitly naming the column index
for each element is not a portable solution, *but* HMIS explicitly
details the correct format for a compliant CSV file, so maybe it's okay.
What do you think, @kfogel?  (I'm looking at lines 1019-1041 of
index.js, here).

POST the data!  It saves correctly in my DB.