Prettify duplicate warning: #25
Move the import button to the bottom of the page, with the other
buttons. Set "import" button to click the file input (the former
"Browse") and then submit the form.  Manage duplicate warning by placing
it in the "results" div for now.  @nttaylor points out that we could
show them a "Possible duplicates detected" link and then, if they
clicked that, give them a popup with more information.