Rewrite import, #25
Read the file in the browser, using the FileReader API.  Check if dates
are in Excel format and convert to YYYY-MM-DD if so (better method
coming soon).  Show duplicate warning -- presently there is an error
where the duplicate warning header may appear even when there are no
duplicates.  Fixing that is the next commit.

More soon to come: removing now-obsolete route and controllers, since we
are handling the file in the browser.  Checking to make sure that the
chosen file is in fact a csv.