Signin/auth works for initial GET request.
Store id_token in the browser, for use in multiple requests (is this
safe?).  Define a new "signIn" route for the Node server to send auth
information to the HMIS server, and use that on user click of sign in
button.  Update the GET controller with auth token.  Pass token along
with initial GET request.

Users can now sign in and fetch the data set for initial searching and
navigation.  Next is to add the token to other API calls, so that users
will be able to create and update clients.  I also need to test this
against a live version of the HMIS API (right now I'm just testing
against my local copy).