Store client ID in local.js: #33
Link to the explanation of how to get a Google client ID for
authentication.  Save this ID in a config file.  Store the generated
client_id in the page and use it for signin.

Since the client-side javascript already needed to have this client_id,
I don't think that embedding it in the page is a problem.