Add login to client
Add a preliminary login mechanism.  The client can now authenticate with
Google and receive an `id_token`, but isn't yet passing it as the value
of the Authorization header in requests to the API, so still won't get
any response.  Next: store the `id_token` in some usual way in the
Javascript, and pass it with each API request.