Authenticate to the API
Pass the id_token with requests to the API, so we can retrieve
information and update clients.  Note that we're not showing anything to
the user that indicates to them whether or not their request to update a
user was successful (see #35).

Next, don't show the search screen until the user has logged in.