Fix export; address error
Change export to use the new return format from the API.  Move
`switchToSearch()` to prevent an error about it not being initialized.
Next is handling the case where the user doesn't have access to the data
or there's some other error more cleanly.

Note: Much of this diff is whitespace, so you might want to view it with
the -w flag.