Show account identity: #36
Show the Google id associated with the logged-in user in the header of
the client app.  This relies on a new endpoint in the HMIS server which
is coming soon, so bear with me (see
https://github.com/PCNI/OpenHMIS/issues/68).  I'll link back to this
commit when that's done.

Note, too, that the clients controller as currently written is
*extremely* redundant, as noted by @kfogel at some point, and needs to
be made more DRY.