Display success/failure messages from import: #25
Show any API errors returned by POST-ing records from the import.  Also
let the user know when items were imported correctly.

Unrelated, but also move the function that gets information about the
logged-in user so that it is only run on page load, not every time we
switch to the search page view.