Send server status to browser
The HMIS server sends some status to the client *server*, which should
be reflected in what the Node server sends to the browser.  Adding the
`.status()` method everywhere should take care of that.

Also: Remove obsolete method for getting information about a user, since
we use the `/externalId` endpoint for that now.  Make a few small text
changes and make sure that the token cookie is used everywhere.