Show import errors to user: #25
If the imported file is not a CSV in the correct format, tell the user
that.  If it is, then show the user the outcome for each line of the
imported file.  The data in each line will either successfully be
POSTed, will be detected as a duplicate, or will fail -- display
messages for each scenario.